Moved algorithm files to API folder
