+++ api/switch_hidden.py
